+++ crates/consensus/src/qbft/component.rs
@@ -8,6 +8,8 @@ use std::{
 
 use futures::future::BoxFuture;
 use k256::{PublicKey, SecretKey};
+use prost::{Message, Name};
+use prost_types::Any;
 use tokio::{sync::mpsc, task::JoinHandle};
 use tokio_util::sync::CancellationToken;
 
@@ -23,7 +25,10 @@ use pluto_core::{
     types::{Duty, DutyType},
 };
 
-use super::{admission, msg, runner};
+use super::{
+    msg::{self, ValueMap},
+    runner,
+};
 
 /// Result returned by outbound QBFT broadcasting.
 pub type BroadcastResult = std::result::Result<(), Box<dyn StdError + Send + Sync + 'static>>;
@@ -92,6 +97,124 @@ pub(crate) enum DecodedValue {
     PriorityResult(pbpriority::PriorityResult),
 }
 
+/// Component result.
+pub type Result<T> = std::result::Result<T, Error>;
+
+/// Component construction and inbound admission errors.
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    /// Peer order did not fit the wire index type.
+    #[error("peer index overflow: {index}")]
+    PeerIndexOverflow {
+        /// Peer order index.
+        index: usize,
+    },
+
+    /// Local peer index is not present in the peer list.
+    #[error("invalid local peer index: {peer_idx}")]
+    InvalidLocalPeerIndex {
+        /// Local peer index.
+        peer_idx: i64,
+    },
+
+    /// Outer consensus message was absent or wrong.
+    #[error("invalid consensus message")]
+    InvalidConsensusMessage,
+
+    /// Inner message type was invalid.
+    #[error("invalid consensus message type")]
+    InvalidConsensusMessageType,
+
+    /// Inner duty type was invalid.
+    #[error("invalid consensus message duty type")]
+    InvalidConsensusMessageDutyType,
+
+    /// Inner round was invalid.
+    #[error("invalid consensus message round")]
+    InvalidConsensusMessageRound,
+
+    /// Inner prepared round was invalid.
+    #[error("invalid consensus message prepared round")]
+    InvalidConsensusMessagePreparedRound,
+
+    /// Message peer index was not in the peer map.
+    #[error("invalid peer index")]
+    InvalidPeerIndex,
+
+    /// Signature verification failed before comparison.
+    #[error("verify consensus message signature: {0}")]
+    VerifyConsensusMessageSignature(#[source] msg::Error),
+
+    /// Signature recovered to a different peer key.
+    #[error("invalid consensus message signature")]
+    InvalidConsensusMessageSignature,
+
+    /// Duty gate rejected the message.
+    #[error("invalid duty")]
+    InvalidDuty,
+
+    /// Justification failed validation.
+    #[error("invalid justification: {0}")]
+    InvalidJustification(#[source] Box<Error>),
+
+    /// Justification duty differed from the outer message duty.
+    #[error("qbft justification duty differs from message duty")]
+    JustificationDutyDiffers,
+
+    /// Inbound Any could not be decoded.
+    #[error("unmarshal any")]
+    UnmarshalAny,
+
+    /// Message wrapper rejected the value map.
+    #[error("{0}")]
+    Msg(#[from] msg::Error),
+
+    /// Duty deadline rejected the message.
+    #[error("duty expired")]
+    DutyExpired,
+
+    /// Receive buffer could not accept the message.
+    #[error("timeout enqueuing receive buffer")]
+    TimeoutEnqueuingReceiveBuffer,
+
+    /// Context was cancelled after expensive verification.
+    #[error("receive cancelled during verification")]
+    ReceiveCancelledDuringVerification,
+}
+
+/// Canonicalizes inbound `Any` values into the hash map used by QBFT messages.
+pub(crate) fn values_by_hash(values: &[Any]) -> Result<ValueMap> {
+    let mut out = ValueMap::new();
+
+    for value in values {
+        let decoded = decode_supported_any(value)?;
+        let hash = match decoded {
+            DecodedValue::UnsignedDataSet(inner) => msg::hash_proto(&inner)?,
+            DecodedValue::PriorityResult(inner) => msg::hash_proto(&inner)?,
+        };
+        out.insert(hash, value.clone());
+    }
+
+    Ok(out)
+}
+
+/// Decodes the protobuf `Any` payload types accepted by this consensus layer.
+pub(crate) fn decode_supported_any(value: &Any) -> Result<DecodedValue> {
+    if value.type_url == pbcore::UnsignedDataSet::type_url() {
+        let decoded = pbcore::UnsignedDataSet::decode(value.value.as_slice())
+            .map_err(|_| Error::UnmarshalAny)?;
+        return Ok(DecodedValue::UnsignedDataSet(decoded));
+    }
+
+    if value.type_url == pbpriority::PriorityResult::type_url() {
+        let decoded = pbpriority::PriorityResult::decode(value.value.as_slice())
+            .map_err(|_| Error::UnmarshalAny)?;
+        return Ok(DecodedValue::PriorityResult(decoded));
+    }
+
+    Err(Error::UnmarshalAny)
+}
+
 pub(crate) enum Subscriber {
     Unsigned(UnsignedSubscriber),
     Priority(PrioritySubscriber),
@@ -148,27 +271,6 @@ pub struct Consensus {
     compare_attestations: bool,
     subscribers: SubscriberSet,
     instances: Mutex<HashMap<Duty, Arc<InstanceIo<msg::Msg>>>>,
-}
-
-/// Component result.
-pub type Result<T> = std::result::Result<T, Error>;
-
-/// Component construction errors.
-#[derive(Debug, thiserror::Error, PartialEq, Eq)]
-pub enum Error {
-    /// Peer order did not fit the wire index type.
-    #[error("peer index overflow: {index}")]
-    PeerIndexOverflow {
-        /// Peer order index.
-        index: usize,
-    },
-
-    /// Local peer index is not present in the peer list.
-    #[error("invalid local peer index: {peer_idx}")]
-    InvalidLocalPeerIndex {
-        /// Local peer index.
-        peer_idx: i64,
-    },
 }
 
 impl Consensus {
@@ -235,40 +337,37 @@ impl Consensus {
         &self,
         ct: &CancellationToken,
         req: Option<pbconsensus::QbftConsensusMsg>,
-    ) -> admission::Result<()> {
-        let pb_msg = req.ok_or(admission::Error::InvalidConsensusMessage)?;
-        let msg = pb_msg
-            .msg
-            .as_ref()
-            .ok_or(admission::Error::InvalidConsensusMessage)?;
+    ) -> Result<()> {
+        let pb_msg = req.ok_or(Error::InvalidConsensusMessage)?;
+        let msg = pb_msg.msg.as_ref().ok_or(Error::InvalidConsensusMessage)?;
 
         self.verify_msg(msg)?;
         let duty = duty_from_msg(msg)?;
 
         if !self.duty_allowed(&duty) {
-            return Err(admission::Error::InvalidDuty);
+            return Err(Error::InvalidDuty);
         }
 
         for justification in &pb_msg.justification {
             self.verify_msg(justification)
-                .map_err(|err| admission::Error::InvalidJustification(Box::new(err)))?;
+                .map_err(|err| Error::InvalidJustification(Box::new(err)))?;
 
             let just_duty = duty_from_msg(justification)
-                .map_err(|err| admission::Error::InvalidJustification(Box::new(err)))?;
+                .map_err(|err| Error::InvalidJustification(Box::new(err)))?;
             if just_duty != duty {
-                return Err(admission::Error::JustificationDutyDiffers);
+                return Err(Error::JustificationDutyDiffers);
             }
         }
 
-        let values = admission::values_by_hash(&pb_msg.values)?;
+        let values = values_by_hash(&pb_msg.values)?;
         let wrapped = msg::Msg::new(msg.clone(), pb_msg.justification.clone(), Arc::new(values))?;
 
         if ct.is_cancelled() {
-            return Err(admission::Error::ReceiveCancelledDuringVerification);
+            return Err(Error::ReceiveCancelledDuringVerification);
         }
 
         if self.add_deadline(duty.clone()).await != AddOutcome::Scheduled {
-            return Err(admission::Error::DutyExpired);
+            return Err(Error::DutyExpired);
         }
 
         let inst = self.get_instance_io(duty);
@@ -280,48 +379,43 @@ impl Consensus {
                     // expires. Its receive task is gone, but late messages
                     // should not abort the sender's broadcast.
                     Err(_) if inst.has_started() => Ok(()),
-                    Err(_) => Err(admission::Error::TimeoutEnqueuingReceiveBuffer),
+                    Err(_) => Err(Error::TimeoutEnqueuingReceiveBuffer),
                 }
             }
-            () = ct.cancelled() => Err(admission::Error::TimeoutEnqueuingReceiveBuffer),
+            () = ct.cancelled() => Err(Error::TimeoutEnqueuingReceiveBuffer),
         }
     }
 
     /// Verifies fields and signature for one raw QBFT message.
-    pub(crate) fn verify_msg(&self, msg: &pbconsensus::QbftMsg) -> admission::Result<()> {
+    pub(crate) fn verify_msg(&self, msg: &pbconsensus::QbftMsg) -> Result<()> {
         if msg.duty.is_none() {
-            return Err(admission::Error::InvalidConsensusMessage);
+            return Err(Error::InvalidConsensusMessage);
         }
 
         if !qbft::MessageType::from_wire(msg.r#type).valid() {
-            return Err(admission::Error::InvalidConsensusMessageType);
+            return Err(Error::InvalidConsensusMessageType);
         }
 
-        let duty = msg
-            .duty
-            .as_ref()
-            .ok_or(admission::Error::InvalidConsensusMessage)?;
-        let duty_type = DutyType::try_from(duty.r#type)
-            .map_err(|_| admission::Error::InvalidConsensusMessageDutyType)?;
+        let duty = msg.duty.as_ref().ok_or(Error::InvalidConsensusMessage)?;
+        let duty_type =
+            DutyType::try_from(duty.r#type).map_err(|_| Error::InvalidConsensusMessageDutyType)?;
         if !duty_type.is_valid() {
-            return Err(admission::Error::InvalidConsensusMessageDutyType);
+            return Err(Error::InvalidConsensusMessageDutyType);
         }
 
         if msg.round <= 0 {
-            return Err(admission::Error::InvalidConsensusMessageRound);
+            return Err(Error::InvalidConsensusMessageRound);
         }
 
         if msg.prepared_round < 0 {
-            return Err(admission::Error::InvalidConsensusMessagePreparedRound);
+            return Err(Error::InvalidConsensusMessagePreparedRound);
         }
 
-        let pubkey = self
-            .pubkey(msg.peer_idx)
-            .ok_or(admission::Error::InvalidPeerIndex)?;
-        let signature_ok = msg::verify_msg_sig(msg, pubkey)
-            .map_err(admission::Error::VerifyConsensusMessageSignature)?;
+        let pubkey = self.pubkey(msg.peer_idx).ok_or(Error::InvalidPeerIndex)?;
+        let signature_ok =
+            msg::verify_msg_sig(msg, pubkey).map_err(Error::VerifyConsensusMessageSignature)?;
         if !signature_ok {
-            return Err(admission::Error::InvalidConsensusMessageSignature);
+            return Err(Error::InvalidConsensusMessageSignature);
         }
 
         Ok(())
@@ -458,26 +552,31 @@ impl Consensus {
 }
 
 /// Extracts the domain duty from a validated raw QBFT message.
-fn duty_from_msg(msg: &pbconsensus::QbftMsg) -> admission::Result<Duty> {
-    let duty = msg
-        .duty
-        .as_ref()
-        .ok_or(admission::Error::InvalidConsensusMessage)?;
-    Duty::try_from(duty).map_err(|_| admission::Error::InvalidConsensusMessageDutyType)
+fn duty_from_msg(msg: &pbconsensus::QbftMsg) -> Result<Duty> {
+    let duty = msg.duty.as_ref().ok_or(Error::InvalidConsensusMessage)?;
+    Duty::try_from(duty).map_err(|_| Error::InvalidConsensusMessageDutyType)
 }
 
 #[cfg(test)]
 pub(crate) mod tests {
     use std::sync::Mutex as StdMutex;
 
+    use prost::{Message, bytes::Bytes};
+    use prost_types::Any;
+    use test_case::test_case;
     use tokio_util::sync::CancellationToken;
 
     use super::*;
     use crate::timer::get_round_timer_func;
     use pluto_core::{
         deadline::{DeadlineCalculator, DeadlinerTask},
+        qbft::SomeMsg,
         types::{DutyType, SlotNumber},
     };
+
+    const REFERENCE_VALUE_HASH: &str =
+        "0a0c0a0430783939120401020304000000000000000000000000000000000000";
+    const REFERENCE_PAYLOAD: &str = "0a6f08021204082a1002200142414cf90756a4241bce7b71e18c6fb9cf91dc96abc6ef1739218974d96e75faf0a15921d47997210232cf064b5e401c6de800fb1f654fcadca0e293dea335fe9242005a200a0c0a04307839391204010203040000000000000000000000000000000000001a440a32747970652e676f6f676c65617069732e636f6d2f636f72652e636f726570622e76312e556e7369676e656444617461536574120e0a0c0a0430783939120401020304";
 
     struct FutureCalculator;
 
@@ -516,7 +615,7 @@ pub(crate) mod tests {
             Err(err) => err,
         };
 
-        assert_eq!(err, Error::InvalidLocalPeerIndex { peer_idx: 3 });
+        assert!(matches!(err, Error::InvalidLocalPeerIndex { peer_idx: 3 }));
     }
 
     #[tokio::test]
@@ -604,6 +703,436 @@ pub(crate) mod tests {
             calls.lock().unwrap().as_slice(),
             ["unsigned-1", "unsigned-2"]
         );
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_invalid_outer_message() {
+        let err = consensus(0, true)
+            .handle(&CancellationToken::new(), None)
+            .await
+            .unwrap_err();
+
+        assert_eq!(err.to_string(), "invalid consensus message");
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_missing_inner_message() {
+        let err = consensus(0, true)
+            .handle(
+                &CancellationToken::new(),
+                Some(pbconsensus::QbftConsensusMsg::default()),
+            )
+            .await
+            .unwrap_err();
+
+        assert_eq!(err.to_string(), "invalid consensus message");
+    }
+
+    #[test_case(|msg: &mut pbconsensus::QbftMsg| msg.r#type = 99, "invalid consensus message type" ; "invalid_message_type")]
+    #[test_case(|msg: &mut pbconsensus::QbftMsg| msg.duty.as_mut().unwrap().r#type = 99, "invalid consensus message duty type" ; "invalid_duty_type")]
+    #[test_case(|msg: &mut pbconsensus::QbftMsg| msg.round = 0, "invalid consensus message round" ; "invalid_round")]
+    #[test_case(|msg: &mut pbconsensus::QbftMsg| msg.prepared_round = -1, "invalid consensus message prepared round" ; "invalid_prepared_round")]
+    #[test_case(|msg: &mut pbconsensus::QbftMsg| msg.peer_idx = 9, "invalid peer index" ; "invalid_peer_idx")]
+    #[tokio::test]
+    async fn verify_msg_rejects_invalid_fields(mutate: fn(&mut pbconsensus::QbftMsg), want: &str) {
+        let consensus = consensus(0, true);
+        let mut msg = signed_msg(0);
+        mutate(&mut msg);
+        if want != "invalid consensus message signature" {
+            msg.signature.clear();
+            msg = sign_for_peer(msg, 0);
+            mutate(&mut msg);
+        }
+
+        let err = consensus.verify_msg(&msg).unwrap_err();
+
+        assert_eq!(err.to_string(), want);
+    }
+
+    #[tokio::test]
+    async fn verify_msg_rejects_missing_duty() {
+        let consensus = consensus(0, true);
+        let mut msg = signed_msg(0);
+        msg.duty = None;
+
+        let err = consensus.verify_msg(&msg).unwrap_err();
+
+        assert_eq!(err.to_string(), "invalid consensus message");
+    }
+
+    #[tokio::test]
+    async fn verify_msg_rejects_empty_signature() {
+        let consensus = consensus(0, true);
+        let mut msg = unsigned_msg(0);
+        msg.signature.clear();
+
+        let err = consensus.verify_msg(&msg).unwrap_err();
+
+        assert_eq!(
+            err.to_string(),
+            "verify consensus message signature: empty signature"
+        );
+    }
+
+    #[tokio::test]
+    async fn verify_msg_rejects_malformed_signature() {
+        let consensus = consensus(0, true);
+        let mut msg = unsigned_msg(0);
+        msg.signature = vec![0x42; 64].into();
+
+        let err = consensus.verify_msg(&msg).unwrap_err();
+
+        assert!(
+            err.to_string()
+                .starts_with("verify consensus message signature: recover pubkey")
+        );
+    }
+
+    #[tokio::test]
+    async fn verify_msg_rejects_wrong_signature() {
+        let consensus = consensus(0, true);
+        let mut msg = unsigned_msg(0);
+        msg.signature = msg::sign_msg(&msg, &secret_key(1)).unwrap().signature;
+        msg.peer_idx = 1;
+
+        let err = consensus.verify_msg(&msg).unwrap_err();
+
+        assert_eq!(err.to_string(), "invalid consensus message signature");
+    }
+
+    #[tokio::test]
+    async fn verify_msg_accepts_valid_signature() {
+        let consensus = consensus(0, true);
+
+        consensus.verify_msg(&signed_msg(0)).unwrap();
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_duty_gate_false() {
+        let err = consensus(0, false)
+            .handle(
+                &CancellationToken::new(),
+                Some(consensus_msg(signed_msg(0))),
+            )
+            .await
+            .unwrap_err();
+
+        assert_eq!(err.to_string(), "invalid duty");
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_invalid_justification() {
+        let mut invalid = signed_msg(0);
+        invalid.round = 0;
+        let outer = pbconsensus::QbftConsensusMsg {
+            msg: Some(signed_msg(0)),
+            justification: vec![invalid],
+            values: vec![],
+        };
+
+        let err = consensus(0, true)
+            .handle(&CancellationToken::new(), Some(outer))
+            .await
+            .unwrap_err();
+
+        assert!(err.to_string().starts_with("invalid justification"));
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_justification_duty_mismatch() {
+        let mut justification = unsigned_msg(0);
+        justification.duty = Some(pbcore::Duty {
+            slot: 43,
+            r#type: i32::try_from(&DutyType::Attester).unwrap(),
+        });
+        let justification = sign_for_peer(justification, 0);
+        let outer = pbconsensus::QbftConsensusMsg {
+            msg: Some(signed_msg(0)),
+            justification: vec![justification],
+            values: vec![],
+        };
+
+        let err = consensus(0, true)
+            .handle(&CancellationToken::new(), Some(outer))
+            .await
+            .unwrap_err();
+
+        assert_eq!(
+            err.to_string(),
+            "qbft justification duty differs from message duty"
+        );
+    }
+
+    #[test]
+    fn values_by_hash_rejects_invalid_type_url() {
+        let err = values_by_hash(&[Any {
+            type_url: "type.googleapis.com/unknown.Type".to_string(),
+            value: vec![],
+        }])
+        .unwrap_err();
+
+        assert_eq!(err.to_string(), "unmarshal any");
+    }
+
+    #[test]
+    fn values_by_hash_rejects_malformed_any_value() {
+        let err = values_by_hash(&[Any {
+            type_url: pbcore::UnsignedDataSet::type_url(),
+            value: b"not-protobuf".to_vec(),
+        }])
+        .unwrap_err();
+
+        assert_eq!(err.to_string(), "unmarshal any");
+    }
+
+    #[test]
+    fn values_by_hash_hashes_decoded_inner_message() {
+        let any = unsigned_any("a", b"first");
+        let values = values_by_hash(std::slice::from_ref(&any)).unwrap();
+        let decoded = pbcore::UnsignedDataSet::decode(any.value.as_slice()).unwrap();
+        let hash = msg::hash_proto(&decoded).unwrap();
+
+        assert_eq!(values.get(&hash), Some(&any));
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_missing_value_hash() {
+        let mut msg = unsigned_msg(0);
+        msg.value_hash = [9u8; 32].to_vec().into();
+        let msg = sign_for_peer(msg, 0);
+
+        let err = consensus(0, true)
+            .handle(&CancellationToken::new(), Some(consensus_msg(msg)))
+            .await
+            .unwrap_err();
+
+        assert_eq!(err.to_string(), "value hash not found in values");
+    }
+
+    #[tokio::test]
+    async fn handle_enqueues_valid_message() {
+        let consensus = consensus(0, true);
+        let any = unsigned_any("a", b"first");
+        let value = pbcore::UnsignedDataSet::decode(any.value.as_slice()).unwrap();
+        let value_hash = msg::hash_proto(&value).unwrap();
+        let mut msg = unsigned_msg(0);
+        msg.value_hash = value_hash.to_vec().into();
+        let msg = sign_for_peer(msg, 0);
+        let duty = duty();
+        let inst = consensus.get_instance_io(duty.clone());
+
+        consensus
+            .handle(
+                &CancellationToken::new(),
+                Some(pbconsensus::QbftConsensusMsg {
+                    msg: Some(msg),
+                    justification: vec![],
+                    values: vec![any],
+                }),
+            )
+            .await
+            .unwrap();
+
+        let mut recv_rx = inst.take_recv_rx().unwrap();
+        let received = recv_rx.try_recv().unwrap();
+        assert_eq!(received.value(), value_hash);
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_deadliner_false_as_duty_expired() {
+        let consensus = Consensus::new(Config {
+            peers: peers(),
+            local_peer_idx: 0,
+            ..config_base(true)
+        })
+        .unwrap();
+
+        let err = consensus
+            .handle(
+                &CancellationToken::new(),
+                Some(consensus_msg(signed_msg(0))),
+            )
+            .await
+            .unwrap_err();
+
+        assert_eq!(err.to_string(), "duty expired");
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_cancellation_after_verification() {
+        let ct = CancellationToken::new();
+        ct.cancel();
+
+        let err = consensus(0, true)
+            .handle(&ct, Some(consensus_msg(signed_msg(0))))
+            .await
+            .unwrap_err();
+
+        assert_eq!(err.to_string(), "receive cancelled during verification");
+    }
+
+    #[tokio::test]
+    async fn handle_waits_for_receive_buffer_capacity() {
+        let consensus = consensus(0, true);
+        let inst = consensus.get_instance_io(duty());
+        let mut recv_rx = inst.take_recv_rx().unwrap();
+        for _ in 0..crate::instance::RECV_BUFFER_SIZE {
+            inst.recv_tx.try_send(wrapped_msg()).unwrap();
+        }
+
+        let ct = CancellationToken::new();
+        let handle = consensus.handle(&ct, Some(consensus_msg(signed_msg(0))));
+        tokio::pin!(handle);
+
+        tokio::select! {
+            result = &mut handle => panic!(
+                "handle completed while receive buffer was full: {result:?}"
+            ),
+            () = tokio::task::yield_now() => {}
+        }
+
+        recv_rx.recv().await.unwrap();
+        tokio::time::timeout(std::time::Duration::from_secs(1), &mut handle)
+            .await
+            .unwrap()
+            .unwrap();
+    }
+
+    #[tokio::test]
+    async fn handle_rejects_full_receive_buffer_after_cancellation() {
+        let consensus = consensus(0, true);
+        let inst = consensus.get_instance_io(duty());
+        let _recv_rx = inst.take_recv_rx().unwrap();
+        for _ in 0..crate::instance::RECV_BUFFER_SIZE {
+            inst.recv_tx.try_send(wrapped_msg()).unwrap();
+        }
+
+        let ct = CancellationToken::new();
+        let handle = consensus.handle(&ct, Some(consensus_msg(signed_msg(0))));
+        tokio::pin!(handle);
+
+        tokio::select! {
+            result = &mut handle => panic!(
+                "handle completed while receive buffer was full: {result:?}"
+            ),
+            () = tokio::task::yield_now() => {}
+        }
+        ct.cancel();
+        let err = tokio::time::timeout(std::time::Duration::from_secs(1), &mut handle)
+            .await
+            .unwrap()
+            .unwrap_err();
+
+        assert_eq!(err.to_string(), "timeout enqueuing receive buffer");
+    }
+
+    #[tokio::test]
+    async fn handle_drops_late_message_after_started_receiver_closed() {
+        let consensus = consensus(0, true);
+        let duty = duty();
+        let inst = consensus.get_instance_io(duty.clone());
+        assert!(inst.maybe_start());
+        drop(inst.take_recv_rx().unwrap());
+        let any = unsigned_any("a", b"first");
+        let value = pbcore::UnsignedDataSet::decode(any.value.as_slice()).unwrap();
+        let value_hash = msg::hash_proto(&value).unwrap();
+        let mut msg = unsigned_msg(0);
+        msg.value_hash = value_hash.to_vec().into();
+        let msg = sign_for_peer(msg, 0);
+
+        consensus
+            .handle(
+                &CancellationToken::new(),
+                Some(pbconsensus::QbftConsensusMsg {
+                    msg: Some(msg),
+                    justification: vec![],
+                    values: vec![any],
+                }),
+            )
+            .await
+            .unwrap();
+
+        assert!(Arc::ptr_eq(&inst, &consensus.get_instance_io(duty)));
+    }
+
+    #[tokio::test]
+    async fn reference_signed_message_is_admitted() {
+        let consensus = consensus(0, true);
+        let mut recv_rx = consensus
+            .get_instance_io(duty())
+            .take_recv_rx()
+            .expect("recv receiver should be available");
+
+        consensus
+            .handle(&CancellationToken::new(), Some(reference_consensus_msg()))
+            .await
+            .expect("reference message should be admitted");
+
+        let received = recv_rx.recv().await.expect("admitted message");
+        assert_eq!(received.source(), 0);
+        assert_eq!(hex::encode(received.value()), REFERENCE_VALUE_HASH);
+        assert_eq!(
+            received.value_source().expect("value source should exist"),
+            reference_any_value()
+        );
+    }
+
+    fn consensus_msg(msg: pbconsensus::QbftMsg) -> pbconsensus::QbftConsensusMsg {
+        pbconsensus::QbftConsensusMsg {
+            msg: Some(msg),
+            justification: vec![],
+            values: vec![],
+        }
+    }
+
+    fn unsigned_msg(peer_idx: i64) -> pbconsensus::QbftMsg {
+        pbconsensus::QbftMsg {
+            r#type: i64::from(qbft::MSG_PRE_PREPARE),
+            duty: Some(pbcore::Duty::try_from(&duty()).unwrap()),
+            peer_idx,
+            round: 1,
+            prepared_round: 0,
+            ..Default::default()
+        }
+    }
+
+    fn signed_msg(peer_idx: i64) -> pbconsensus::QbftMsg {
+        sign_for_peer(unsigned_msg(peer_idx), peer_idx)
+    }
+
+    fn sign_for_peer(msg: pbconsensus::QbftMsg, peer_idx: i64) -> pbconsensus::QbftMsg {
+        let seed = u8::try_from(peer_idx.checked_add(1).unwrap()).unwrap();
+        msg::sign_msg(&msg, &secret_key(seed)).unwrap()
+    }
+
+    fn unsigned_any(key: &str, value: &'static [u8]) -> Any {
+        Any::from_msg(&pbcore::UnsignedDataSet {
+            set: [(key.to_string(), Bytes::from_static(value))].into(),
+        })
+        .unwrap()
+    }
+
+    fn reference_consensus_msg() -> pbconsensus::QbftConsensusMsg {
+        pbconsensus::QbftConsensusMsg::decode(
+            hex::decode(REFERENCE_PAYLOAD)
+                .expect("valid fixture hex")
+                .as_slice(),
+        )
+        .expect("reference payload should decode")
+    }
+
+    fn reference_value() -> pbcore::UnsignedDataSet {
+        let mut set = std::collections::BTreeMap::new();
+        set.insert("0x99".to_string(), Bytes::from_static(&[1, 2, 3, 4]));
+        pbcore::UnsignedDataSet { set }
+    }
+
+    fn reference_any_value() -> Any {
+        Any::from_msg(&reference_value()).expect("value should pack")
+    }
+
+    fn wrapped_msg() -> msg::Msg {
+        msg::Msg::new(unsigned_msg(0), vec![], Arc::default()).unwrap()
     }
 
     pub(crate) fn consensus(local_peer_idx: i64, duty_allowed: bool) -> Consensus {
