+++ crates/consensus/src/qbft/mod.rs
@@ -1,12 +1,11 @@
 //! QBFT consensus wrapper.
 
-mod admission;
 mod component;
 pub(crate) mod definition;
 pub(crate) mod runner;
 
 pub use component::{
-    BroadcastResult, Broadcaster, Config, Consensus, DutyGater, Error, Peer, SnifferSink,
+    BroadcastResult, Broadcaster, Config, Consensus, DutyGater, Error, Peer, Result, SnifferSink,
     SubscriberResult,
 };
 pub use runner::{Error as RunnerError, Result as RunnerResult};
@@ -20,7 +19,5 @@ pub mod p2p;
 pub(crate) mod sniffer;
 pub(crate) mod transport;
 
-#[cfg(test)]
-mod interop_test;
 #[cfg(test)]
 mod qbft_run_test;
